Receiving data from GNURadio bug fix and other fixes for bugs causing the release builds to not work

+++ SDRReradiation/DeviceReceivers.h
@@ -54,7 +54,7 @@ class DeviceReceivers
 
 		static uint32_t maxCorrelationCount;
 		uint8_t count;
-		uint8_t initializedDevices;
+		int initializedDevices;
 		FrequencyRange frequencyRange;
 
 		#ifdef _WIN32
